+++ src/main/java/gtclassic/block/GTBlockMachine.java
@@ -43,6 +43,7 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
@@ -83,9 +84,13 @@ public class GTBlockMachine extends GTBlockMultiID {
 
 	@Override
 	public void addInformation(ItemStack stack, World worldIn, List<String> tooltip, ITooltipFlag flagIn) {
+		if (this == GTBlocks.tileBlastFurnace) {
+			tooltip.add(TextFormatting.RED + I18n.format("WARNING THIS TILE WILL BE MOVED TO GTC EXPANSION NEXT RELEASE"));
+		}
 		for (int i = 0; i < this.size; i++) {
 			tooltip.add(I18n.format(this.getUnlocalizedName().replace("tile", "tooltip") + i));
 		}
+		
 	}
 
 	@Override