+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 - Added original GT1 helium fusion recipe, costs 40 million EU but will output 1,048,576 EU into adjacent Supercondensator(must be front or back of the fusion tile).
 
 # 1.04 (Not Released)
+- WARNING! In the next version (1.0.5) the Blast Furnace will become part of GTC Expansion (by Trinsdar)
+- WARNING! In the next version (1.0.5) GTC fluid blocks and molten fluids will be removed, molten fluids will be part of GTC Expansion as well.
+- Added temporary recipes for getting your resources back from the Reinforced casing/block and Blast Furnace before next release!
+- Added more crafting recipes for duct tape, that hopefully indicate what its capible of doing in world : )
 - Removed clicking behavior from quantum chest, but i redid the input/ouput logic to be instantaneous.
-- Removed fluid container clicking behavior from magic energy converter to avoid a dupe bug.
+- Removed Supercontainer as it has no use, and no future use. Different from the Supercondensator and Superconductor.
+- Fixed recipes that use the superconductor item, now use ore dictionary for other mods.
 - Fixed Translocator duplicating items when wrenched.
 - Fixed Charge-o-mat not charging baubles slots.
 - Finally got around to making progress on the basic workbench, can now keep inventory in crafting slots - limit one active user as a time.